Rename network package to coroutine for coroutine-related utilities
