+++ core/common/src/main/kotlin/com/google/samples/apps/nowinandroid/core/coroutine/NiaDispatchers.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.nowinandroid.core.network
+package com.google.samples.apps.nowinandroid.core.coroutine
 
 import javax.inject.Qualifier
 import kotlin.annotation.AnnotationRetention.RUNTIME
